save the images in the gallery using media and some visual aid

+++ src/app/[name].tsx
@@ -1,16 +1,38 @@
-import { View, Image } from "react-native";
+import { View, Image, Alert } from "react-native";
 import { router, useLocalSearchParams } from "expo-router";
 import * as FileSystem from "expo-file-system/legacy";
 import { MaterialIcons } from "@expo/vector-icons";
 import { Stack } from "expo-router";
+import { useState } from "react";
+import * as MediaLibray from "expo-media-library";
 
 const ImageScreen = () => {
   const { name } = useLocalSearchParams<{ name: string }>();
   const fullUri = FileSystem.documentDirectory + name;
+  const [permissionResponse, requestPermission] = MediaLibray.usePermissions();
+  const [isSaving, setIsSaving] = useState(false);
 
   const onDelete = async () => {
-    await FileSystem.deleteAsync(fullUri);
-    router.back();
+    Alert.alert("Delete Media", "Are you sure you want to delete this Media?", [
+      { text: "Cancel", style: "cancel" },
+      {
+        text: "Delete",
+        style: "destructive",
+        onPress: async () => {
+          await FileSystem.deleteAsync(fullUri);
+          router.back();
+        },
+      },
+    ]);
+  };
+
+  const onSave = async () => {
+    if (permissionResponse?.status !== "granted") {
+      await requestPermission();
+    }
+    setIsSaving(true);
+    await MediaLibray.createAssetAsync(fullUri);
+    setIsSaving(false);
   };
 
   return (
@@ -27,8 +49,8 @@ const ImageScreen = () => {
                 color="crimson"
               />
               <MaterialIcons
-                onPress={() => {}}
-                name="save"
+                onPress={onSave}
+                name={isSaving ? "hourglass-empty" : "save"}
                 size={26}
                 color="dimgray"
               />
